requirements.txt
# Core data science libraries
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Statistical modeling
statsmodels>=0.14.0
lifelines>=0.27.0

# Machine learning
xgboost>=1.7.0
lightgbm>=4.0.0
catboost>=1.2.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Marketing attribution specific
networkx>=3.1.0  # For Markov chain models
itertools-recipes>=1.0.0  # For Shapley value combinations

# Data processing
pyarrow>=12.0.0  # For parquet files
sqlalchemy>=2.0.0  # Database connections

# API integrations
requests>=2.31.0
google-ads>=21.0.0
facebook-business>=17.0.0
google-analytics-data>=0.17.0

# Configuration and utilities
python-dotenv>=1.0.0
pyyaml>=6.0
click>=8.0.0  # CLI interface

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Dashboard and web
streamlit>=1.25.0
dash>=2.14.0
flask>=2.3.0

# Time series and forecasting
prophet>=1.1.4
pmdarima>=2.0.3

# Privacy and compliance
cryptography>=41.0.0
hashlib-compat>=1.0.0